fully modular and consistent api with added lightweight testing, bug fixes, and quality of life changes
