+++ signal_memory_engine_v1/.env-example
@@ -1,43 +1,18 @@
-# =========================
-# Core API Keys (required)
-# =========================
-OPENAI_API_KEY=sk-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
-PINECONE_API_KEY=pc-xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
+# === Pinecone ===
+PINECONE_API_KEY=<your-pinecone-api-key>
+PINECONE_ENVIRONMENT=<your-pinecone-environment>   # e.g. us-west1-gcp
+PINECONE_INDEX=<your-pinecone-index-name>          # defaults to "signal-engine" if unset
 
-# =========================
-# Pinecone config
-# =========================
-PINECONE_ENV=us-west1-gcp
+# === OpenAI ===
+OPENAI_API_KEY=<your-openai-api-key>
+OPENAI_MODEL=<your-openai-model>                   # e.g. gpt-4o-mini
 
-# =========================
-# SQLite storage
-# =========================
-# Path to the local SQLite DB used by /signal and /drift endpoints.
-SME_DB_PATH=./data/signal.db
+# === MLflow ===
+MLFLOW_TRACKING_URI=                               # optional; falls back to ./mlruns
+MLFLOW_EXPERIMENT_NAME=<your-experiment-name>      # defaults to "signal-memory-engine"
 
-# =========================
-# MLflow tracking
-# =========================
-# If empty, the app falls back to a local file:// URI under ./mlruns (Windows-safe).
-MLFLOW_TRACKING_URI=
-MLFLOW_EXPERIMENT_NAME=signal-memory-engine
+# === Local DB (Signals) ===
+SME_DB_PATH=<your-SME_DB_PATH>                     # defaults to ./data/signal.db if unset
 
-# =========================
-# Optional model overrides
-# (the app has defaults; set these only if you want to change them)
-# =========================
-LLM_MODEL=gpt-3.5-turbo
-EMBED_MODEL=sentence-transformers/all-MiniLM-L6-v2
-OPENAI_EMBED_MODEL=text-embedding-ada-002
-
-# =========================
-# Server (optional)
-# =========================
-APP_HOST=127.0.0.1
-APP_PORT=8000
-
-# =========================
-# Testing
-# =========================
-# Set to 1 during unit tests to skip initializing external services.
-SME_TEST_MODE=0
+# === Testing ===
+SME_TEST_MODE=0                                    # optional test switch; default 0 (off)
