+++ signal_memory_engine_v1/agents/axis_agent.py
@@ -1,28 +1,21 @@
 # agents/axis_agent.py
 import os
+from functools import lru_cache
 from dotenv import load_dotenv
 from scripts.langchain_retrieval import build_qa_chain
 
-# Load .env so keys/vars are available
 load_dotenv()
 
-# Env normalization
-PINECONE_ENV = os.getenv("PINECONE_ENVIRONMENT", "us-west1-gcp")
-OPENAI_MODEL = os.getenv("OPENAI_MODEL", "gpt-4o-mini")
-
-# Environment-specific index name for Axis™
-INDEX_NAME = os.getenv("PINECONE_INDEX_AXIS", "axis-memory")
-
-# Build the QA chain and vectorstore for Axis™
-qa_axis, store_axis = build_qa_chain(
-    pinecone_api_key=os.getenv("PINECONE_API_KEY"),
-    pinecone_env=PINECONE_ENV,
-    index_name=INDEX_NAME,
-    openai_api_key=os.getenv("OPENAI_API_KEY"),
-    embed_model="sentence-transformers/all-MiniLM-L6-v2",
-    llm_model=OPENAI_MODEL,
-    k=3,
-)
-
-# Human‐readable role for Axis™
 ROLE_AXIS = "Axis™ Relationship Architect"
+
+@lru_cache(maxsize=1)
+def get_axis_chain():
+    return build_qa_chain(
+        pinecone_api_key=os.getenv("PINECONE_API_KEY"),
+        pinecone_env=os.getenv("PINECONE_ENVIRONMENT", "us-west1-gcp"),
+        index_name=os.getenv("PINECONE_INDEX_AXIS", "axis-memory"),
+        openai_api_key=os.getenv("OPENAI_API_KEY"),
+        embed_model="sentence-transformers/all-MiniLM-L6-v2",
+        llm_model=os.getenv("OPENAI_MODEL", "gpt-4o-mini"),
+        k=3,
+    )
