+++ signal_memory_engine_v1/mlruns/176452795178630947/traces/tr-0ad9988e739d2bc22c552fed70f155aa/request_metadata/mlflow.source.name
@@ -1,0 +1,1 @@
+C:\Users\nate\Desktop\coherence_protocol\Signal_memory_engine\signal_memory_engine_v1\venv\Scripts\uvicorn
