+++ signal_memory_engine_v1/mlruns/176452795178630947/traces/tr-0ad9988e739d2bc22c552fed70f155aa/request_metadata/mlflow.trace.sizeStats
@@ -1,0 +1,1 @@
+{"total_size_bytes": 10303, "num_spans": 6, "max": 3098, "p25": 764, "p50": 1045, "p75": 1955}
