+++ signal_memory_engine_v1/mlruns/176452795178630947/traces/tr-0ad9988e739d2bc22c552fed70f155aa/request_metadata/mlflow.trace.tokenUsage
@@ -1,0 +1,1 @@
+{"input_tokens": 97, "output_tokens": 4, "total_tokens": 101}
