+++ signal_memory_engine_v1/mlruns/176452795178630947/traces/tr-0ad9988e739d2bc22c552fed70f155aa/request_metadata/mlflow.traceInputs
@@ -1,0 +1,1 @@
+{"query": "Current biometric readings:\n• HRV: 91.8 ms\n• Temp: 37.2 °C\n• Blink rate: 15.6 bpm\n\nUse this context to answer the question below.\n\nQuestion: What is emotional recursion?"}
