+++ signal_memory_engine_v1/tests/test_coherence_commons.py
@@ -1,3 +1,5 @@
+# tests/test_coherence_commons.py
+
 """
 Coherence stage tests (lightweight & offline)
