+++ signal_memory_engine_v1/tests/test_storage_sqlite_store.py
@@ -1,3 +1,5 @@
+# tests/test_storage_sqlite_store.py
+
 """
 Storage layer tests (sqlite_store)
